Fix next.js title generation by allowing the metadata API to work by not having the whole main page render on the client.
