+++ project/zemn.me/app/layout.tsx
@@ -1,5 +1,8 @@
-import ClientLayout from 'project/zemn.me/app/ClientLayout';
+import 'project/zemn.me/app/base.css';
+
+import { Metadata } from 'next/types';
 import { ReactNode } from 'react';
+import { HeaderTagsAppRouter } from 'ts/next.js';
 
 export interface Props {
 	readonly children?: ReactNode;
@@ -9,10 +12,6 @@ export function RootLayout({ children }: Props) {
 	return (
 		<html>
 			<head>
-				<meta
-					content="Personal website and profile of Thomas Neil James Shadwell, also known as Zemnmez."
-					name="description"
-				/>
 				<link href="https://fonts.googleapis.com" rel="preconnect" />
 				<link
 					crossOrigin="anonymous"
@@ -23,12 +22,24 @@ export function RootLayout({ children }: Props) {
 					href="https://fonts.googleapis.com/css2?family=Lora:ital,wght@0,400;0,700;1,400;1,700&display=swap"
 					rel="stylesheet"
 				/>
+				<link href="/icon.svg" rel="icon" type="image/svg+xml" />
+				<link
+					href="/icon.svg"
+					rel="apple-touch-icon"
+					type="image/svg+xml"
+				/>
+				<HeaderTagsAppRouter />
 			</head>
-			<body>
-				<ClientLayout>{children}</ClientLayout>
-			</body>
+			<body>{children}</body>
 		</html>
 	);
 }
 
 export default RootLayout;
+
+export const metadata: Metadata = {
+	metadataBase: new URL('https://zemn.me'),
+	twitter: {
+		creator: '@zemnmez',
+	},
+};
