+++ project/zemn.me/app/page.tsx
@@ -1,5 +1,3 @@
-'use client';
-import Head from 'next/head';
 import { Metadata } from 'next/types';
 import style from 'project/zemn.me/app/style.module.css';
 import * as kenwood from 'project/zemn.me/assets/kenwood';
@@ -37,16 +35,6 @@ function LetterHead() {
 export default function Main() {
 	return (
 		<main className={style.main}>
-			<Head>
-				<title lang={lang.get(bio.Bio.who.handle)}>
-					{lang.text(bio.Bio.who.handle)}
-				</title>
-				<meta
-					content="zemn.me git https://github.com/zemnmez/go.git"
-					name="go-import"
-				/>
-			</Head>
-
 			<video
 				autoPlay
 				className={style.headerBgv}
@@ -244,8 +232,5 @@ export const metadata: Metadata = {
 		firstName: lang.text(bio.Bio.who.firstName),
 		lastName: lang.text(bio.Bio.who.lastName),
 		username: lang.text(bio.Bio.who.handle),
-	},
-	twitter: {
-		creator: '@zemnmez',
 	},
 };
