+++ project/zemn.me/components/timeline/timeline.tsx
@@ -1,3 +1,4 @@
+'use client';
 import Immutable from 'immutable';
 import memoizee from 'memoizee';
 import * as Bio from 'project/zemn.me/bio';
@@ -213,6 +214,8 @@ function Year({
 		Immutable.List<Bio.Event>
 	>;
 }) {
+	const firstItem = months.first()?.first();
+	const fullYear = firstItem?.date.getFullYear() ?? 0;
 	return (
 		<div className={style.year}>
 			<div className={style.yearIndicator} lang={year.get('language')}>
@@ -220,10 +223,7 @@ function Year({
 			</div>
 			{/*↓ unicode for 'no specified language, roman numerals used for numbering'*/}
 			<div className={style.ageIndicator} lang="zxx-u-nu-romanlow">
-				{romanize(
-					(months.first(undefined)?.first()?.date.getFullYear() ??
-						0) - 1994
-				)}
+				{romanize(fullYear - 1994)}
 			</div>
 
 			<div className={style.content}>
