+++ ts/pulumi/shadwell.im/lucy/pages/_app.tsx
@@ -1,10 +1,10 @@
 import type { AppProps } from 'next/app';
-import { HeaderTags } from 'ts/next.js';
+import { HeaderTagsPagesRouter } from 'ts/next.js';
 
 export function App({ Component, pageProps }: AppProps) {
 	return (
 		<>
-			<HeaderTags />
+			<HeaderTagsPagesRouter />
 			<Component {...pageProps} />
 		</>
 	);
