+++ ts/pulumi/shadwell.im/luke/pages/_app.tsx
@@ -1,10 +1,10 @@
 import type { AppProps } from 'next/app';
-import { HeaderTags } from 'ts/next.js';
+import { HeaderTagsAppRouter } from 'ts/next.js';
 
 export function App({ Component, pageProps }: AppProps) {
 	return (
 		<>
-			<HeaderTags />
+			<HeaderTagsAppRouter />
 			<Component {...pageProps} />
 		</>
 	);
